Refactored the exception dialog
Made the exception class private
Added two public functions to display the dialog. One blocking and one not.
Moved the import
