+++ src/amulet_editor/models/widgets/__init__.py
@@ -4,4 +4,3 @@ from ._icon import AStylableSvgWidget, AIconButton, ATooltipIconButton
 from ._label import QElidedLabel
 from ._menu import QMenuWidget
 from ._line import QHLine
-from ._traceback_dialog import AmuletTracebackDialog, DisplayException
