+++ src/amulet_editor/models/widgets/traceback_dialog/_traceback_dialog.ui
